35174 Can't edit Transactions - Fixed Scientific Names cell, should now display scientific name correctly
