+++ packages/dina-ui/components/index.ts
@@ -72,3 +72,4 @@ export * from "./collection/material-sample/RestrictionField";
 export * from "./page/PageLayout";
 export * from "./seqdb/ngs-index/NgsIndexField";
 export * from "./workbook";
+export * from "./collection/material-sample/organismUtils";
